Added code to see Leslie's approach
It's not exactly how Leslie approached the problem. Instead of taking one example of each class I've taken examples at random.  Training time is 1/4th the time (7.8 seconds vs 30.8 seconds on Colab) it would take on a normal network. Using this approach I got a test accuracy of about 96% (training the entire dataset is giving an accuracy of 97%).